requirements.txt
sentence-transformers
langchain
langchain-huggingface
langchain-community
flask
pypdf
python-dotenv
tiktoken
chromadb
-e .